# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Urological Review System API
  description: Backend API for managing urological residency application reviews
  version: 1.0.0
  contact:
    name: Urological Review System
    email: support@example.com
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.yourproductiondomain.com
    description: Production server
paths:
...
  /api/applicants:
    get:
      summary: Get all applicants
      description: Retrieve all applicants in the system
      responses:
        '200':
          description: Successfully retrieved applicants
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DatabaseApplicant'
    post:
      summary: Create a new applicant
      description: Add a new applicant to the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicantRequest'
      responses:
        '201':
          description: Applicant created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/DatabaseApplicant'
        '400':
          description: Validation error
        '409':
          description: Applicant already exists
...
  /api/applicants/{id}:
    get:
      summary: Get applicant by ID
      description: Retrieve a specific applicant by their UUID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved applicant
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/DatabaseApplicant'
        '404':
          description: Applicant not found
    put:
      summary: Update applicant
      description: Update an existing applicant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateApplicantRequest'
      responses:
        '200':
          description: Applicant updated successfully
        '404':
          description: Applicant not found
    delete:
      summary: Delete applicant
      description: Remove an applicant from the system
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Applicant deleted successfully
        '404':
          description: Applicant not found
...
  /api/reviews/applicant/{applicantId}:
    get:
      summary: Get reviews for applicant
      description: Retrieve all reviews for a specific applicant
      parameters:
        - name: applicantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved reviews
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DatabaseReview'
...
components:
  schemas:
    ApplicantCategory:
      type: string
      enum: [regular, i-sub]
...
    DecisionType:
      type: string
      enum: ["Definitely Interview", Maybe, "Do Not Interview"]
    DatabaseApplicant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        external_id:
          type: string
        name:
          type: string
        category:
          $ref: '#/components/schemas/ApplicantCategory'
        details:
          type: string
        site_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DatabaseReview:
      type: object
      properties:
        id:
          type: string
          format: uuid
        applicant_id:
          type: string
          format: uuid
        reviewer_name:
          type: string
        preference:
          type: integer
          minimum: 1
          maximum: 5
        pressure:
          type: integer
          minimum: 1
          maximum: 5
        underserved:
          type: integer
          minimum: 1
          maximum: 5
        leadership:
          type: integer
          minimum: 1
          maximum: 5
        academic:
          type: integer
          minimum: 1
          maximum: 5
        research:
          type: integer
          minimum: 1
          maximum: 5
        personal:
          type: integer
          minimum: 1
          maximum: 5
        notes:
          type: string
        decision:
          $ref: '#/components/schemas/DecisionType'
        total_score:
          type: integer
        site_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CreateApplicantRequest:
      type: object
      required:
        - external_id
        - name
      properties:
        external_id:
          type: string
        name:
          type: string
        category:
          $ref: '#/components/schemas/ApplicantCategory'
        details:
          type: string
    UpdateApplicantRequest:
      type: object
      properties:
        name:
          type: string
        category:
          $ref: '#/components/schemas/ApplicantCategory'
        details:
          type: string
...